+++ CHANGELOG.md
@@ -5,3 +5,8 @@
 ## 1.0.1
 
 * readme changed.
+
+## 1.0.2
+
+* Changed class Name from Email_OTP to EmailOTP.
+* otpLength property added.